+++ page.php
@@ -3,10 +3,16 @@
 <div id="primary" class="content-area">
 	<main id="main" class="site-main">
 		<?php
-		// Inizia il loop di WordPress
-		while ( have_posts() ) : the_post();
-			the_content();
-		endwhile;
+			$slug = get_post_field('post_name', get_post());
+			$template_path = locate_template('pages/page-' . $slug . '.php');
+
+			if ( $template_path ) {
+				require $template_path;
+			} else {
+				while ( have_posts() ) : the_post();
+					the_content();
+				endwhile;
+			}
 		?>
 	</main>
 </div>