+++ parts/separator.php
@@ -1,0 +1,1 @@
+<div class="separator">*&nbsp;&nbsp;&nbsp;*&nbsp;&nbsp;&nbsp;*</div>